--- src/main/scala/tech/cryptonomic/cloud/nautilus/model/ApiKey.scala
@@ -1,14 +1,0 @@
-package tech.cryptonomic.cloud.nautilus.model
-
-import java.sql.Timestamp
-
-/** Class representing API Key */
-case class ApiKey(
-    keyId: Int,
-    key: String,
-    resourceId: Int,
-    userId: Int,
-    tierId: Int,
-    dateIssued: Option[Timestamp],
-    dateSuspended: Option[Timestamp]
-)
