--- src/main/scala/tech/cryptonomic/cloud/nautilus/model/DoobieConfig.scala
@@ -1,4 +1,0 @@
-package tech.cryptonomic.cloud.nautilus.model
-
-/** Class representing Doobie configuration */
-case class DoobieConfig(url: String, user: String, password: String)
