--- src/main/scala/tech/cryptonomic/cloud/nautilus/model/HttpConfig.scala
@@ -1,4 +1,0 @@
-package tech.cryptonomic.cloud.nautilus.model
-
-/** Represents http server configuration */
-case class HttpConfig(host: String, port: Int)
