--- src/main/scala/tech/cryptonomic/cloud/nautilus/model/UsageLeft.scala
@@ -1,4 +1,0 @@
-package tech.cryptonomic.cloud.nautilus.model
-
-/** Example representation of the usage per key */
-case class UsageLeft(key: String, daily: Int, monthly: Int)
