--- src/main/scala/tech/cryptonomic/cloud/nautilus/model/User.scala
@@ -1,14 +1,0 @@
-package tech.cryptonomic.cloud.nautilus.model
-
-import java.sql.Timestamp
-
-/** Class representing User */
-case class User(
-    userId: Int,
-    userName: String,
-    userEmail: String,
-    userRole: String,
-    registrationDate: Timestamp,
-    accountSource: Option[String],
-    accountDescription: Option[String]
-)
