--- src/main/scala/tech/cryptonomic/cloud/nautilus/model/UserWithoutId.scala
@@ -1,13 +1,0 @@
-package tech.cryptonomic.cloud.nautilus.model
-
-import java.sql.Timestamp
-
-/** Class used in user registration and update */
-case class UserWithoutId(
-    userName: String,
-    userEmail: String,
-    userRole: String,
-    registrationDate: Timestamp,
-    accountSource: Option[String],
-    accountDescription: Option[String]
-)
