--- src/main/scala/tech/cryptonomic/cloud/nautilus/repositories/UserRepo.scala
@@ -1,18 +1,0 @@
-package tech.cryptonomic.cloud.nautilus.repositories
-
-import scala.language.higherKinds
-
-import tech.cryptonomic.cloud.nautilus.model.{User, UserWithoutId}
-
-/** Trait representing User repo queries */
-trait UserRepo[F[_]] {
-
-  /** Creates user */
-  def createUser(userReg: UserWithoutId): F[Int]
-
-  /** Updates user */
-  def updateUser(user: User): F[Unit]
-
-  /** Returns user */
-  def getUser(userId: Int): F[Option[User]]
-}
