--- src/main/scala/tech/cryptonomic/cloud/nautilus/routes/endpoint/EndpointsUtil.scala
@@ -1,18 +1,0 @@
-package tech.cryptonomic.cloud.nautilus.routes.endpoint
-
-import endpoints.algebra
-import endpoints.algebra.Documentation
-
-/** Utility trait for extending Endpoints functionality */
-trait EndpointsUtil extends algebra.Responses {
-
-  /** Extension for using Created status code */
-  def created[A](response: Response[A], invalidDocs: Documentation): Response[A]
-
-  /** Extensions for [[Response]]. */
-  implicit class CustomResponseExtensions[A](response: Response[A]) {
-
-    /** syntax for `created` */
-    final def withCreatedStatus(notFoundDocs: Documentation = None): Response[A] = created(response, notFoundDocs)
-  }
-}
