+++ src/main/scala/tech/cryptonomic/nautilus/cloud/model/DoobieConfig.scala
@@ -1,0 +1,4 @@
+package tech.cryptonomic.nautilus.cloud.model
+
+/** Class representing Doobie configuration */
+case class DoobieConfig(url: String, user: String, password: String)
