+++ src/main/scala/tech/cryptonomic/nautilus/cloud/model/HttpConfig.scala
@@ -1,0 +1,4 @@
+package tech.cryptonomic.nautilus.cloud.model
+
+/** Represents http server configuration */
+case class HttpConfig(host: String, port: Int)
