+++ src/main/scala/tech/cryptonomic/nautilus/cloud/model/UsageLeft.scala
@@ -1,0 +1,4 @@
+package tech.cryptonomic.nautilus.cloud.model
+
+/** Example representation of the usage per key */
+case class UsageLeft(key: String, daily: Int, monthly: Int)
