+++ src/main/scala/tech/cryptonomic/nautilus/cloud/model/UserWithoutId.scala
@@ -1,0 +1,13 @@
+package tech.cryptonomic.nautilus.cloud.model
+
+import java.sql.Timestamp
+
+/** Class used in user registration and update */
+case class UserWithoutId(
+    userName: String,
+    userEmail: String,
+    userRole: String,
+    registrationDate: Timestamp,
+    accountSource: Option[String],
+    accountDescription: Option[String]
+)
