+++ src/main/scala/tech/cryptonomic/nautilus/cloud/repositories/UserRepo.scala
@@ -1,0 +1,17 @@
+package tech.cryptonomic.nautilus.cloud.repositories
+
+import scala.language.higherKinds
+import tech.cryptonomic.nautilus.cloud.model.{User, UserWithoutId}
+
+/** Trait representing User repo queries */
+trait UserRepo[F[_]] {
+
+  /** Creates user */
+  def createUser(userReg: UserWithoutId): F[Int]
+
+  /** Updates user */
+  def updateUser(user: User): F[Unit]
+
+  /** Returns user */
+  def getUser(userId: Int): F[Option[User]]
+}
