+++ src/main/scala/tech/cryptonomic/nautilus/cloud/services/ApiKeyService.scala
@@ -1,0 +1,15 @@
+package tech.cryptonomic.nautilus.cloud.services
+
+import tech.cryptonomic.nautilus.cloud.model.ApiKey
+
+import scala.language.higherKinds
+
+/** API keys service */
+trait ApiKeyService[F[_]] {
+
+  /** Returns all API keys from the DB */
+  def getAllApiKeys: F[List[ApiKey]]
+
+  /** Checks if API key is valid */
+  def validateApiKey(apiKey: String): F[Boolean]
+}
